combine reference columns in view
